Refactor item management by separating actions into ItemRowActions component and implementing updateItem service
